requirements.txt
azure-identity
azure-mgmt-apimanagement
azure-mgmt-resource
python-dotenv
py-healthcheck==1.10.1
azure-functions